+++ packages/legend-application/src/index.ts
@@ -51,5 +51,4 @@ export * from './stores/ApplicationStoreTestUtils';
 
 // value specification
 export * from './stores/ValueSpecificationModifierHelper';
-export * from './stores/ValueSpecificationUtils';
 export * from './components/BasicValueSpecificationEditor';
