+++ packages/legend-query/src/stores/filterOperators/QueryBuilderFilterOperator_GreaterThan.ts
@@ -26,13 +26,13 @@ import {
   PRIMITIVE_TYPE,
   isSuperType,
   SUPPORTED_FUNCTIONS,
+  buildPrimitiveInstanceValue,
 } from '@finos/legend-graph';
 import { UnsupportedOperationError } from '@finos/legend-shared';
 import {
   buildFilterConditionState,
   buildFilterConditionExpression,
 } from './QueryBuilderFilterOperatorHelper';
-import { buildPrimitiveInstanceValue } from '@finos/legend-application';
 import { QUERY_BUILDER_SUPPORTED_FUNCTIONS } from '../../QueryBuilder_Const';
 import { generateDefaultValueForPrimitiveType } from '../QueryBuilderValueSpecificationBuilderHelper';
 import { getNonCollectionValueSpecificationType } from '../QueryBuilderOperatorsHelper';
